Fix voc and initial column headers

+++ src/read_sgp40.h
@@ -2,6 +2,6 @@
 #define READ_SGP40
 
 void init_sgp40();
-uint8_t read_sgp40();
+uint32_t read_sgp40();
 
 #endif
